Revert "refactor test data generation"
This reverts commit b07321bb0b237542d980c10d78829958c1a6e1ce.
